+++ src/main/java/utec/proyectofinal/Proyecto/Final/UTEC/dtos/response/CultivarDTO.java
@@ -5,7 +5,7 @@ import utec.proyectofinal.Proyecto.Final.UTEC.business.entities.Especie;
 
 @Data
 public class CultivarDTO {
-    private Integer cultivarID;
+    private Long cultivarID;
     private String especie;
     private String nombre;
 }
